// src/SqlEditor/sqlUtils/type.d.ts
import { AST } from 'node-sql-parser';
import { SqlErrorTypeEnum } from './enum';

export interface ValidateResult {
  validatorName: string;
  level: Level;
  result: boolean;
  message: string;
}
export interface ValidateSqlResult {
  message: string;
  uiMessages: string | string[];
  sqlParseResult: SqlParseResult | null;
  sqlErrorType: SqlErrorTypeEnum;
  validateResults: ValidateResult[];
}
export interface ValidateAstParams extends ValidatorFuncProps {
  validators: ValidatorConfig[];
}
export interface ValidatorFuncProps {
  extractedAstList?: AST[];
  originTableList?: string[];
  invalidMessage?: string;
}
export type Level = 'warn' | 'error' | 'pass';
export type ValidatorFunc = ({
  extractedAstList,
  invalidMessage,
}: ValidatorFuncProps) => { validateResult: boolean; invalidMessage: string }; // Return: false(warn or error)； true(pass)
export interface ValidatorConfig {
  validate: ValidatorFunc;
  level?: Level; // default: warn
  invalidMessage?: string;
}

export interface SqlParseResult {
  fullAst: AST | AST[];
  fullTableList: string[];
  fullColumnList: string[];
  sqlDataList: {
    ast: any;
    sql: string;
    extractedAstList: AST[];
    tableList: string[];
    columnList: string[];
  }[];
}

interface ValidatorFuncProps {
  extractedAstList?: AST[];
  originTableList?: string[];
}
export interface ValidatorConfig {
  validate: ValidatorFunc;
  level?: Level; // default: warn
}